src/propythia/DNA/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
fonttools==4.30.0
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
requests==2.27.1
six==1.16.0
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
typing_extensions==4.1.1
urllib3==1.26.8
biopython==1.79
